Lion Random Fact Function
Like the Tiger class, we'll now add a function that will generate a random fact for our Lion. We'll highlight the ability to access a property in a class by testing the property with an if statement.

This function returns a String instance, randomFact. We'll set the value of the randomFact variable by testing if the property isAlphaMale is true or not.